Move some constants to the shared integ test library
